+++ app/src/main/java/com/example/senamit/stationaryhutpro/adapters/CategoryProductAdapter.java
@@ -2,7 +2,6 @@ package com.example.senamit.stationaryhutpro.adapters;
 
 import android.content.Context;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -46,7 +45,7 @@ public class CategoryProductAdapter extends RecyclerView.Adapter<CategoryProduct
         if (product!= null){
             holder.txtProductNumber.setText(product.get(position).getProductNumber());
             holder.txtProductName.setText(product.get(position).getProductName());
-            holder.txtProductPrice.setText(Integer.toString(product.get(position).getProductPrice()));
+            holder.txtProductPrice.setText(Double.toString(product.get(position).getProductPrice()));
             String imageUrl = product.get(position).getImageUrl();
             Picasso.with(context).load(imageUrl).into(holder.imageProduct);
 //            Log.i(TAG, "the position is "+position);
